// src/app/models/perfil-associado.model.ts
export interface PerfilAssociado {
	id: number;
	associadoId: number;
	fotoUrl: string;
	nomeProfissional: string;
	nomeEmpresa: string;
	logomarcaUrl: string;
	exibirEnderecoComercial: boolean;
	telefonePrincipal: string;
	telefoneSecundario: string;
	emailPrincipal: string;
	site: string;
	linkedin: string;
	instagram: string;
	youtube: string;
	outraRedeSocial: string;
	oQueEuFaco: string;
	publicoIdeal: string;
	principalProblemaQueResolvo: string;
	conexoesEstrategicas: string;
	interessesPessoais: string;
	clusterNome: string;
	atuacaoEspecificaNome: string;
	equipeNome: string;
	status: string;
	dataIngresso: string;
	vencimento: string;
}